+++ examples/example_dialog.rs
@@ -30,7 +30,7 @@ impl actix::Handler<woab::Signal> for WindowActor {
     fn handle(&mut self, msg: woab::Signal, ctx: &mut Self::Context) -> Self::Result {
         Ok(match msg.name() {
             "close" => {
-                gtk4::main_quit();
+                // gtk4::main_quit();
                 None
             }
             "open_dialog" => {
@@ -66,30 +66,30 @@ impl actix::Handler<woab::Signal> for WindowActor {
                 None
             }
             "reset" => {
-                ctx.spawn(
-                    async move {
-                        woab::run_dialog(
-                            &gtk4::MessageDialog::new::<gtk4::ApplicationWindow>(
-                                None,
-                                gtk4::DialogFlags::all(),
-                                gtk4::MessageType::Question,
-                                gtk4::ButtonsType::YesNo,
-                                "Reset the counters?",
-                            ),
-                            true,
-                        )
-                        .await
-                    }
-                    .into_actor(self)
-                    .map(|response, actor, _ctx| {
-                        if response == gtk4::ResponseType::Yes {
-                            actor.yes_count = 0;
-                            actor.no_count = 0;
-                            actor.widgets.yes_count.set_text("0");
-                            actor.widgets.no_count.set_text("0");
-                        }
-                    }),
-                );
+                // ctx.spawn(
+                    // async move {
+                        // woab::run_dialog(
+                            // &gtk4::MessageDialog::new::<gtk4::ApplicationWindow>(
+                                // None,
+                                // gtk4::DialogFlags::all(),
+                                // gtk4::MessageType::Question,
+                                // gtk4::ButtonsType::YesNo,
+                                // "Reset the counters?",
+                            // ),
+                            // true,
+                        // )
+                        // .await
+                    // }
+                    // .into_actor(self)
+                    // .map(|response, actor, _ctx| {
+                        // if response == gtk4::ResponseType::Yes {
+                            // actor.yes_count = 0;
+                            // actor.no_count = 0;
+                            // actor.widgets.yes_count.set_text("0");
+                            // actor.widgets.no_count.set_text("0");
+                        // }
+                    // }),
+                // );
                 None
             }
             _ => msg.cant_handle()?,
@@ -173,7 +173,7 @@ fn main() -> anyhow::Result<()> {
             })
     });
 
-    gtk4::main();
-
+    // gtk4::main();
+    woab::close_actix_runtime()??;
     Ok(())
 }