+++ examples/example_taggged_signals.rs
@@ -30,10 +30,6 @@ impl actix::Actor for WindowActor {
         self.widgets.win_app.show();
         self.recalculate();
     }
-
-    fn stopped(&mut self, _ctx: &mut Self::Context) {
-        gtk4::main_quit();
-    }
 }
 
 impl actix::Handler<woab::Signal> for WindowActor {
@@ -42,7 +38,7 @@ impl actix::Handler<woab::Signal> for WindowActor {
     fn handle(&mut self, msg: woab::Signal, ctx: &mut Self::Context) -> Self::Result {
         Ok(match msg.name() {
             "close" => {
-                gtk4::main_quit();
+                // gtk4::main_quit();
                 None
             }
             "click_button" => {
@@ -52,7 +48,7 @@ impl actix::Handler<woab::Signal> for WindowActor {
                     .row_addend
                     .instantiate()
                     .with_object("row_addend", |row_addend| {
-                        self.widgets.lst_addition.add(&row_addend);
+                        // self.widgets.lst_addition.add(&row_addend);
                         self.addends.insert(addend_id, (row_addend, Some(0)));
                     })
                     .connect_to((addend_id, ctx.address()));
@@ -74,16 +70,16 @@ impl actix::Handler<woab::Signal<usize>> for WindowActor {
     fn handle(&mut self, msg: woab::Signal<usize>, _ctx: &mut Self::Context) -> Self::Result {
         Ok(match msg.name() {
             "addend_changed" => {
-                let woab::params!(buffer: gtk4::TextBuffer) = msg.params()?;
-                let new_number: Option<isize> = buffer
-                    .text(&buffer.start_iter(), &buffer.end_iter(), true)
-                    .and_then(|text| text.parse().ok());
-                if let Some((_, number)) = self.addends.get_mut(msg.tag()) {
-                    if new_number != *number {
-                        *number = new_number;
-                        self.recalculate();
-                    }
-                }
+                // let woab::params!(buffer: gtk4::TextBuffer) = msg.params()?;
+                // let new_number: Option<isize> = buffer
+                    // .text(&buffer.start_iter(), &buffer.end_iter(), true)
+                    // .and_then(|text| text.parse().ok());
+                // if let Some((_, number)) = self.addends.get_mut(msg.tag()) {
+                    // if new_number != *number {
+                        // *number = new_number;
+                        // self.recalculate();
+                    // }
+                // }
                 None
             }
             "remove_addend" => {
@@ -127,6 +123,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         });
     });
 
-    gtk4::main();
+    // gtk4::main();
+    woab::close_actix_runtime()??;
     Ok(())
 }