+++ CHANGELOG.md
@@ -5,7 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## 0.8.0 - 2023-08-19
 ### Changed
+- [**BREAKING**] Upgraded gtk-rs to version 0.18. This is a breaking change
+  because gtk-rs' API had some changes in this release:
+  - Instead of `gtk::Inhibit`, use `glib::Propagation`. `Inhibit(false)` should
+    become `Propagation::Stop` and `Inhibit(true)` should become
+    `Propagation::Proceed`.
+
+## 0.7.0 - 2022-09-06
+### Changed
+- [**BREAKING**] Upgraded gtk-rs to version 0.15. This is a breaking change
+  because gtk-rs' API had some changes in this release.
+- Updated Actix version to 0.13.
+- **BREAKING** `woab::close_actix_runtime` must be called after `gtk::main()` now.
 - `woab::close_actix_runtime` return an error instead of panicing if the
   runtime is closed or in use. This is breaking because now it is returnes two
   nested `Result`s, but it's a minor function that's not used all over the